Moving iree/base/file_path.h to iree/base/internal/.
