+++ iree/hal/local/loaders/BUILD
@@ -33,11 +33,10 @@ cc_library(
     deps = [
         "//iree/base:api",
         "//iree/base:dynamic_library",
-        "//iree/base:file_path",
         "//iree/base:flatcc",
         "//iree/base:tracing",
         "//iree/base/internal:file_io",
-        "//iree/base:tracing",
+        "//iree/base/internal:file_path",
         "//iree/hal:api",
         "//iree/hal/local",
         "//iree/schemas:dylib_executable_def_c_fbs",
