+++ iree/hal/local/loaders/legacy_library_loader.cc
@@ -15,8 +15,8 @@
 #include "iree/hal/local/loaders/legacy_library_loader.h"
 
 #include "iree/base/dynamic_library.h"
-#include "iree/base/file_path.h"
 #include "iree/base/internal/file_io.h"
+#include "iree/base/internal/file_path.h"
 #include "iree/base/tracing.h"
 #include "iree/hal/local/local_executable.h"
